Moved everything to one test
